Modified README.md; Moved content of amami/parsers/main_partser.py to amami/parsers/__init__.py to get rid of complexity of having an additional file. Now all the high-level classes and functions for the creation of the parsers are in the amami/parsers/__init__.py file.
